fix: Ensure mapeo always closes and show closing window while processes are being destroyed. Fixes #398
