+++ static/closing.html
@@ -43,7 +43,7 @@
   <body>
     <div class="root">
       <h1>
-        Esperando que termina sincronización... (5 minutos)
+        Esperando que termina Mapeo... 
       </h1>
     </div>
   </body>
